fix: Move @types/prettier to dependencies
./src/index.ts imports the 'Options' type from @types/prettier as
'PrettierOptions', and re-exports it as part of the package signature.
Because of this, any TypeScript project that installs
json-schema-to-typescript as a dependency or devDependency (e.g. to run
it programmatically) will attempt to locate @types/prettier--and will
run into trouble, because @types/prettier is not installed.

To avoid this, @types/prettier should also be pulled in as a dependency
of json-schema-to-typescript.

This fixes a possible oversight made in:
1a91ef007cd7262331e81edc6a809777f5f38e4f